// frontend/src/generated/soundcloud.ts
export interface paths {
    "/me": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** Returns the authenticated user’s information. */
        get: {
            parameters: {
                query?: never;
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                200: components["responses"]["Me"];
                401: components["responses"]["Unauthorized"];
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/resolve": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** Resolves soundcloud.com and on.soundcloud.com URLs to Resource URLs to use with the API. */
        get: {
            parameters: {
                query: {
                    /**
                     * @description SoundCloud URL
                     * @example https://soundcloud.com/user-434241656
                     */
                    url: components["parameters"]["url"];
                };
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                302: components["responses"]["Found"];
                404: components["responses"]["NotFound"];
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
// ...
        User: {
            /** @description URL to a JPEG image */
            avatar_url?: string;
            /** @description city */
            city?: string;
            /** @description country */
            country?: string;
            /** @description description */
            description?: string;
            /** @description discogs name */
            discogs_name?: string;
            /** @description first name */
            first_name?: string;
            /** @description number of followers */
            followers_count?: number;
            /** @description number of followed users */
            followings_count?: number;
            /** @description first and last name */
            full_name?: string;
            /** @description unique identifier */
            urn?: string;
            /** @description kind of resource */
            kind?: string;
            /**
             * Format: date-time
             * @description profile creation datetime
             */
            created_at?: string;
            /**
             * Format: date-time
             * @description last modified datetime
             */
            last_modified?: string;
            /** @description last name */
            last_name?: string;
            /** @description permalink of the resource */
            permalink?: string;
            /** @description URL to the SoundCloud.com page */
            permalink_url?: string;
            /** @description subscription plan of the user */
            plan?: string;
            /** @description number of public playlists */
            playlist_count?: number;
            /** @description number of favorited public tracks */
            public_favorites_count?: number;
            /** @description number of reposts from user */
            reposts_count?: number;
            /** @description number of public tracks */
            track_count?: number;
            /** @description API resource URL */
            uri?: string;
            /** @description username */
            username?: string;
            /** @description a URL to the website */
            website?: string;
            /** @description a custom title for the website */
            website_title?: string;
            /**
             * @deprecated
             * @description a list subscriptions associated with the user
             */
            subscriptions?: unknown;
        } | null;
// ...
        Track: {
            /** @description Track title. */
            title?: string;
            /** @description URL to a JPEG image. */
            artwork_url?: string;
            /** @description Tempo. */
            bpm?: number;
            /** @description Number of comments. */
            comment_count?: number;
            /** @description Is commentable. */
            commentable?: boolean;
            /** @description Created timestamp. */
            created_at?: string;
            /** @description Track description. */
            description?: string;
            /** @description NUmber of downloads. */
            download_count?: number;
            /** @description Is downloadable. */
            downloadable?: string;
            /** @description Track duration. */
            duration?: number;
            /**
             * @deprecated
             * @description Embeddable by.
             */
            embeddable_by?: string;
            /** @description Number of favoritings. */
            favoritings_count?: number;
            /** @description Genre */
            genre?: string;
            /** @description Track identifier. */
            urn?: string;
            /** @description ISRC code. */
            isrc?: string;
            /** @description Key signature. */
            key_signature?: string;
            /** @description Type of object (track). */
            kind?: string;
            /** @description Label user name. */
            label_name?: string;
            /** @description License */
            license?: string;
            /** @description Optional artist name, when different from user. */
            metadata_artist?: string;
            /** @description Permalink URL. */
            permalink_url?: string;
            /** @description Number of plays. */
            playback_count?: number;
            /** @description Purchase title. */
            purchase_title?: string;
            /** @description Purchase URL. */
            purchase_url?: string;
            /** @description Release. */
            release?: string;
            /** @description Day of release. */
            release_day?: number;
            /** @description Month of release. */
            release_month?: number;
            /** @description Year of release. */
            release_year?: number;
            /** @description Type of sharing (public/private). */
            sharing?: string;
            /**
             * @deprecated
             * @description URL to stream (preview only)
             */
            stream_url?: string;
            /** @description Is streamable. */
            streamable?: boolean;
            /** @description Tags. */
            tag_list?: string;
            /** @description Track URI. */
            uri?: string;
            user?: components["schemas"]["User"] & unknown;
            /** @description Is user's favourite. It is only set when fetching search results or single track, otherwise it is false. */
            user_favorite?: boolean;
            /** @description Number of plays by a user. */
            user_playback_count?: number;
            /** @description Waveform URL. */
            waveform_url?: string;
            /** @description List of countries where track is available. */
            available_country_codes?: string;
            /**
             * @description Level of access the user (logged in or anonymous) has to the track.
             *       * `playable` - user is allowed to listen to a full track.
             *       * `preview` - user is allowed to preview a track, meaning a snippet is available
             *       * `blocked` - user can only see the metadata of a track, no streaming is possible
             * @enum {string|null}
             */
            access?: "playable" | "preview" | "blocked" | null;
            /** @description URL to download a track. */
            download_url?: string;
            /** @description Number of reposts. */
            reposts_count?: number;
            /** @description Secret URL. */
            secret_uri?: string;
        };
        Tracks: {
            collection?: components["schemas"]["Track"][];
            next_href?: string;
        };
        /** @deprecated */
        TracksList: components["schemas"]["Track"][];
        /** @description Soundcloud Playlist Object */
        Playlist: {
            /** @description Playlist title. */
            title?: string;
            /** @description Playlist identifier. */
            urn?: string;
            /** @description Type of Soundcloud object (playlist). */
            kind?: string;
            /** @description URL to a JPEG image. */
            artwork_url?: string;
            /** @description Created timestamp. */
            created_at?: string;
            /** @description Playlist description. */
            description?: string;
            /** @description is downloadable. */
            downloadable?: boolean;
            /** @description Playlist duration. */
            duration?: number;
            /** @description European Article Number. */
            ean?: string;
            /** @description Embeddable by. */
            embeddable_by?: string;
            /** @description Playlist genre. */
            genre?: string;
            /** @description Label user identifier. */
            label_id?: number;
            /** @description Label name. */
            label_name?: string;
            /** @description Last modified timestamp. */
            last_modified?: string;
            /** @description License. */
            license?: string;
            /** @description Playlist permalink. */
            permalink?: string;
            /** @description Playlist permalink URL. */
            permalink_url?: string;
            /** @description Type of playlist. */
            playlist_type?: string;
            /** @description Purchase title. */
            purchase_title?: string;
            /** @description Purchase URL. */
            purchase_url?: string;
            /** @description Release. */
            release?: string;
            /** @description Day of release. */
            release_day?: number;
            /** @description Month of release. */
            release_month?: number;
            /** @description Year of release. */
            release_year?: number;
            /** @description Type of sharing (private/public). */
            sharing?: string;
            /** @description Is streamable. */
            streamable?: boolean;
            /** @description Tags. */
            tag_list?: string;
            /** @description Count of tracks. */
            track_count?: number;
            /** @description List of tracks. */
            tracks?: components["schemas"]["Track"][];
            /** @description Playlist type. */
            type?: string;
            /** @description Playlist URI. */
            uri?: string;
            user?: components["schemas"]["User"];
            /** @description User identifier. */
            user_urn?: string;
            /** @description Count of playlist likes. */
            likes_count?: number;
            label?: (Record<string, never> | null) & components["schemas"]["User"];
            /** @description tracks URI. */
            tracks_uri?: string | null;
            /** @description Tags. */
            tags?: string | null;
        };
// ...
    };
// ...
}